server/requirements.txt
# Flask web framework
Flask>=2.3.0
flask-cors>=4.0.0

# Image processing (flexible versions for Windows compatibility)
Pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0

# Machine Learning - YOLO (flexible versions)
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0

# Additional dependencies
PyYAML>=6.0
matplotlib>=3.5.0
scipy>=1.9.0
psutil>=5.8.0
pandas>=1.5.0